// src/types/UserProfile.ts
export type UserProfile = {
  firstName: string | null;
  lastName: string | null;
  email: string | null;
  phone: string | null;
  address: string | null;
  city: string | null;
  state: string | null;
  zipCode: string | null;
  dateOfBirth: string | null;
  medicalHistory: string | null;
  allergies: string | null;
  medications: string | null;
};